requirements.txt
Flask==2.3.2
flask-cors==4.0.0
mysql-connector-python==8.0.33
Werkzeug==2.3.6
